+++ MFA/src/com/example/mfa/gamepanel/MainThread.java
@@ -153,7 +153,7 @@ public class MainThread extends Thread {
 
 		if (statusIntervalTimer >= lastStatusStore + STAT_INTERVAL) {
 			// calculate the actual frames pers status check interval
-			double actualFps = (double)(frameCountPerStatCycle / (STAT_INTERVAL / 1000));
+			double actualFps = frameCountPerStatCycle / (STAT_INTERVAL / 1000);
 
 			//stores the latest fps in the array
 			fpsStore[(int) statsCount % FPS_HISTORY_NR] = actualFps;
